Move ProjectCard to components folder
