+++ front/app/containers/Admin/inspirationHub/PinnedProjects/index.tsx
@@ -7,9 +7,8 @@ import useProjectLibraryProjects from 'api/project_library_projects/useProjectLi
 
 import { FormattedMessage } from 'utils/cl-intl';
 
+import ProjectCard from '../components/ProjectCard/ProjectCard';
 import messages from '../messages';
-
-import ProjectCard from './ProjectCard';
 
 const PinnedProjects = () => {
   const { data: appConfiguration } = useAppConfiguration();
