+++ front/app/containers/Admin/inspirationHub/components/ProjectCard/CardImage.tsx
